+++ packages/html/src/orgchart/orgchart.spec.tsx
@@ -1,5 +1,4 @@
-import { classNames, kendoThemeMaps } from '../misc';
-import { OrgchartNode } from '../orgchart';
+import { classNames } from '../misc';
 
 export const ORGCHART_CLASSNAME = `k-orgchart`;
 
@@ -8,17 +7,13 @@ const states = [];
 const options = {};
 
 export type KendoOrgchartProps = {
-    children?: JSX.Element | JSX.Element[];
     width?: string;
     height?: string;
-    orientation?: 'horizontal' | 'vertical';
-    justifyContent?: null | 'start' | 'center' | 'end' | 'stretch' | 'around';
 };
 
 const defaultProps = {
     width: '100%',
     height: '300px',
-    orientation: 'horizontal'
 };
 
 export const Orgchart = (
@@ -26,34 +21,10 @@ export const Orgchart = (
         React.HTMLAttributes<HTMLDivElement>
 ) => {
     const {
-        children,
         width = defaultProps.width,
         height = defaultProps.height,
-        orientation = defaultProps.orientation,
-        justifyContent,
         ...other
     } = props;
-
-    const chartNodes : JSX.Element[] = [];
-    const chartGroups : JSX.Element[] = [];
-
-    if (children) {
-        if (Array.isArray(children)) {
-            children.map((child) => {
-                if (child.type === OrgchartNode) {
-                    chartNodes.push( child );
-                } else {
-                    chartGroups.push( child );
-                }
-            });
-        } else {
-            if (children.type === OrgchartNode) {
-                chartNodes.push( children );
-            } else {
-                chartGroups.push( children );
-            }
-        }
-    }
 
     return (
         <div
@@ -64,31 +35,7 @@ export const Orgchart = (
             )}>
 
             <div className="k-orgchart-container" style={{ width: width, height: height }}>
-                <div
-                    className={classNames(
-                        'k-orgchart-group',
-                        'k-pos-absolute',
-                        {
-                            [`k-${kendoThemeMaps.orientationMap[orientation!] || orientation}`]: orientation,
-                            [`k-justify-content-${justifyContent}`]: justifyContent,
-                        }
-                    )}
-                    style={{ width: '100%' }}>
-
-                    {chartNodes.length > 0 &&
-                                            <div
-                                                className={classNames(
-                                                    'k-orgchart-node-container',
-                                                    'k-justify-content-around',
-                                                    {
-                                                        [`k-${kendoThemeMaps.orientationMap[orientation!] || orientation}`]: orientation,
-                                                    }
-                                                )}>
-                                                <>{chartNodes}</>
-                                            </div>
-                    }
-                    <>{chartGroups}</>
-                </div>
+                {props.children}
             </div>
         </div>
     );