+++ packages/html/src/orgchart/tests/orgchart-group-states.tsx
@@ -1,4 +1,4 @@
-import { Orgchart, OrgchartGroup, OrgchartNode } from '../../orgchart';
+import { OrgchartGroup, OrgchartGroupVertical, OrgchartNode, OrgchartNodeGroup } from '../../orgchart';
 
 
 const styles = `
@@ -12,28 +12,20 @@ export default () =>(
         <style>{styles}</style>
         <div id="test-area" className="k-d-grid k-grid-cols-2">
 
-            <span>Default</span>
-            <span>Focus</span>
-
-            <section>
-                <Orgchart width="100%" height="431px" justifyContent="around">
-                    <OrgchartGroup title="Group Title" subtitle="Group SubTitle" orientation="vertical">
-                        <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
-                        <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
-                        <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
-                    </OrgchartGroup>
-                </Orgchart>
-            </section>
-
-            <section>
-                <Orgchart width="100%" height="431px" justifyContent="around">
-                    <OrgchartGroup title="Group Title" subtitle="Group SubTitle" orientation="vertical" focus>
-                        <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
-                        <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
-                        <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
-                    </OrgchartGroup>
-                </Orgchart>
-            </section>
+            {[ "normal", ...OrgchartNodeGroup.states ].map((state) => (
+                <section>
+                    <span>{state}</span>
+                    <OrgchartGroupVertical width="100%" height="431px">
+                        <OrgchartGroup justifyContent="around">
+                            <OrgchartNodeGroup title="Group Title" subtitle="Group SubTitle" orientation="vertical" { ...{ [state]: true } }>
+                                <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
+                                <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
+                                <OrgchartNode title="Card Title" subtitle="Card, Subtitle" details avatar="DK" avatarType="text" color="#5bc0de"></OrgchartNode>
+                            </OrgchartNodeGroup>
+                        </OrgchartGroup>
+                    </OrgchartGroupVertical>
+                </section>
+            ))}
         </div>
     </>
 );