+++ packages/classic/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.3-dev.3](https://github.com/telerik/kendo-themes/compare/v7.0.3-dev.2...v7.0.3-dev.3) (2023-12-04)
+
+**Note:** Version bump only for package @progress/kendo-theme-classic
+
+
+
+
+
+## [7.0.3-dev.2](https://github.com/telerik/kendo-themes/compare/v7.0.3-dev.1...v7.0.3-dev.2) (2023-11-27)
+
+**Note:** Version bump only for package @progress/kendo-theme-classic
+
+
+
+
+
 ## [7.0.3-dev.1](https://github.com/telerik/kendo-themes/compare/v7.0.3-dev.0...v7.0.3-dev.1) (2023-11-20)
 
 **Note:** Version bump only for package @progress/kendo-theme-classic